+++ functor-monad/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Revision history for functor-monad
 
+## 0.1.1.1 -- 2024-09-29
+
+* Add a module `TrailAsAdjunction` to the examples directory,
+  just as an explanation of theoretical background.
+
+* Fix some documentation errors
+
+## 0.1.1.0 -- 2024-02-07
+
+* The first published version
+
 ## 0.1.0.0 (review) -- 2023-12-23
 
 * First version